+++ SamSmithNZ2017/SamSmithNZ2017.DB/dbo/Tables/TabTrack_old.sql
@@ -4,7 +4,7 @@
     [TrackPath]  VARCHAR (100) NULL,
     [TrackOrder] INT           NULL,
     [AlbumKey]   INT           NULL,
-    [Rating]     SMALLINT      NULL,
+    [Rating]     INT      NULL,
     CONSTRAINT [PK_Tracks] PRIMARY KEY CLUSTERED ([TrackKey] ASC)
 );
