+++ SamSmithNZ2017/SamSmithNZ2017.DB/dbo/Tables/ffl_tracking.sql
@@ -1,7 +1,7 @@
 ﻿CREATE TABLE [dbo].[ffl_tracking] (
     [show_date]               DATETIME NULL,
-    [show_key]                SMALLINT NULL,
-    [ffl_show_key]            SMALLINT NULL,
+    [show_key]                INT NULL,
+    [ffl_show_key]            INT NULL,
     [show_location_different] BIT      NULL,
     [show_setlist_different]  BIT      NULL
 );
